Simplified bot functions for higher success rate
Previously, the bot tried, but was unable to choose a good card for a given enemy prediction. Since I couldn't find an algorithm that allows me to predict whether it would be better to under-, or overbid the opponent or whether it would be a good idea to deliberately lose a round, the method now uses a more static bias-based method to recognize an opponent that plays a middle field strat as well to adjust accordingly.